+++ csatar-plugins/csatar/updates/builder_table_update_csatar_csatar_scouts_5.php
@@ -1,0 +1,23 @@
+<?php namespace Csatar\Csatar\Updates;
+
+use Schema;
+use October\Rain\Database\Updates\Migration;
+
+class BuilderTableUpdateCsatarCsatarScouts5 extends Migration
+{
+    public function up()
+    {
+        Schema::table('csatar_csatar_scouts', function($table)
+        {
+            $table->boolean('is_active')->default(0)->change();
+        });
+    }
+    
+    public function down()
+    {
+        Schema::table('csatar_csatar_scouts', function($table)
+        {
+            $table->boolean('is_active')->default(1)->change();
+        });
+    }
+}